Moved setContinuousProfilesSampleRate into ExperimentalOptions
