+++ sentry/src/test/java/io/sentry/SentryOptionsTest.kt
@@ -239,7 +239,7 @@ class SentryOptionsTest {
     @Test
     fun `when continuousProfilesSampleRate is set to a 0, isProfilingEnabled is false and isContinuousProfilingEnabled is false`() {
         val options = SentryOptions().apply {
-            this.continuousProfilesSampleRate = 0.0
+            this.experimental.continuousProfilesSampleRate = 0.0
         }
         assertFalse(options.isProfilingEnabled)
         assertFalse(options.isContinuousProfilingEnabled)
@@ -266,19 +266,19 @@ class SentryOptionsTest {
     @Test
     fun `when setContinuousProfilesSampleRate is set to exactly 0, value is set`() {
         val options = SentryOptions().apply {
-            this.continuousProfilesSampleRate = 0.0
+            this.experimental.continuousProfilesSampleRate = 0.0
         }
         assertEquals(0.0, options.continuousProfilesSampleRate)
     }
 
     @Test
     fun `when setContinuousProfilesSampleRate is set to higher than 1_0, setter throws`() {
-        assertFailsWith<IllegalArgumentException> { SentryOptions().continuousProfilesSampleRate = 1.0000000000001 }
+        assertFailsWith<IllegalArgumentException> { SentryOptions().experimental.continuousProfilesSampleRate = 1.0000000000001 }
     }
 
     @Test
     fun `when setContinuousProfilesSampleRate is set to lower than 0, setter throws`() {
-        assertFailsWith<IllegalArgumentException> { SentryOptions().continuousProfilesSampleRate = -0.0000000000001 }
+        assertFailsWith<IllegalArgumentException> { SentryOptions().experimental.continuousProfilesSampleRate = -0.0000000000001 }
     }
 
     @Test
@@ -607,7 +607,7 @@ class SentryOptionsTest {
     fun `when profiling is disabled, isEnableAppStartProfiling is always false`() {
         val options = SentryOptions()
         options.isEnableAppStartProfiling = true
-        options.continuousProfilesSampleRate = 0.0
+        options.experimental.continuousProfilesSampleRate = 0.0
         assertFalse(options.isEnableAppStartProfiling)
     }
 
@@ -615,7 +615,7 @@ class SentryOptionsTest {
     fun `when setEnableAppStartProfiling is called and continuous profiling is enabled, isEnableAppStartProfiling is true`() {
         val options = SentryOptions()
         options.isEnableAppStartProfiling = true
-        options.continuousProfilesSampleRate = 1.0
+        options.experimental.continuousProfilesSampleRate = 1.0
         assertTrue(options.isEnableAppStartProfiling)
     }
